+++ tests/cases/tabs/test_prepare.py
@@ -1,14 +1,15 @@
 import glob
 import os
 
-from pymol import cmd
 import pytest
+from pymol import cmd
 
 from REvoDesign.tools.customized_widgets import set_widget_value
 
 from ...conftest import TestWorker
 
 os.environ["PYTEST_QT_API"] = "pyqt5"
+
 
 @pytest.mark.order(2)
 class TestREvoDesignPlugin_TabPrepare: